Update installation.rst
From the man page of usermod for the option -G: "If the user is currently a member of a group which is not listed, the user will be removed from the group. This behaviour can be changed via the -a option, which appends the user to the current supplementary group list." In my case the usermod command without the -a removed me from all other groups I was in and set my gropus to e.g. vboxusers.