Introduce a number of changes:
Add extra pretty-printing
Introduce a "KILL" event for the analyzer, so that it's now aware of attempts by a monitored process to terminate another monitored process or itself.  Later we could enable the process dumper to dump the exit state of these tasks, but currently cuckoomon just uses the event notification to flush its logs prior to the termination actually taking place
Make all the process injection code aware of whether the thread/process currently being injected into is suspended or not -- this way we won't delay execution unnecessarily due to multiple NtMapViewOfSection into a suspended process
Don't bother with the explicit apc argument to inject anymore, it will be inferred based on the information available
Make sure we sleep a second after creating the mutex that the various monitored processes are waiting for.  Since the analyzer itself is creating the mutex, cuckoomon is only periodically checking for the existence of the mutex.  If we immediately shut down the processes, they may not have had a chance to react to the creation of the shutdown mutex.

These changes require an updated loader and cuckoomon, though some backward compatibility has been left in.