The existing PDF analysis and peepdf itself both need a lot of work, this is just some small initial changes with lots more being worked on until it's ready for release.
The whole object listing was inaccurate, it was labeling as "encoded" any type of non-stream object, and any non-javascript streams had useless information tagged onto them.  Instead of this, simply display all this information as "Javascript Streams".  Likewise a key named "Streams" was incorrect, as it contained simple keywords (correctly named by pdfid) which could be simple syntax elements or names used as dictionary keys.  Add in reporting of author, creator, and producer of the PDF.

Some deficiencies noted already:
* peepdf doesn't appear to have an easy built-in value walker, perhaps with some recursion detection.  PDF references can occur in many places within a PDF and peepdf currently appears to require manual dereferencing (eg. for dictionaries where an element's value is a reference) which seems to also mess with some of peepdf's assumptions about PDF layouts (eg. that simple strings will appear for certain values instead of references to string objects)
* peepdf doesn't handle modern malicious PDFs involving actions on clicking on the document.  Specifically, link type annotations with an action (/A) of a URI is unhandled.  As is any proper handling of URIs in parsing the PDF -- it can currently only extract URLs via regex parses in javascript.  Code will be added in the future to parse out the Base URI from the Catalog dict and prepend it to partial URIs used throughout the document.