Add HTTP info to getiocs API and multiple bugfixes
Fix a bug where the target file name could have unicode, preventing json from being able to parse it. Also add in some checks for dict keys when dealing with network data, as these sometimes may not exist. Also fix a bug for when the machine key doesn't exist in the info dict when listing tasks.