explicitly state that the http.status is a string
this is to avoid error in elasticsearch mapping.
http.status is declared as long type in the elasticsearch index template
since it's originally an int.
however, if you encounter a http.status that is "None", this would cause
a mapping error since "None" is a string .
so just declare them both as string.