Allow malheur to automatically delete bad reports
This prevents log spam for every analysis when we have issues with specific reports. The warning message configured to be deleted is often times caused by broken analyses or ctrl-c'ing out of reporting early. Warnings can be added in the future if need be, this is just the only one I've observed.