Fix a very old bug introduced by: https://github.com/cuckoosandbox/cuckoo/commit/8fcad5c42ac278cf714021bf00e2bbfb2253e48d
As the timeout applies not just to the initial connect but to any
operation occurring on the socket thereafter, it was observed especially
with large files that the timeout would be hit and the upload prematurely
and silently truncated.  Now all uploads will proceed correctly up to the
maximum allowed size.